Removed while loops because it messes up lots of stuff.
According to the FTA people, you should never have while or for loops in
your code. The lock the staff out of being able to control the robot.
This might get rid of the weird RobotDrive errors on the Driver Station.
Auton is now controlled by a step integer and two switch-case
statements.